Renames the project to watchful
The name disrupt-o-meter was maybe not as optional as we would have
liked, and therefore this commit renames the entire project to watchful
