+++ internal/watchful/cfg/cfg_suite_test.go
@@ -18,43 +18,16 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package dom_test
+package cfg_test
 
 import (
-	cfclient "github.com/cloudfoundry-community/go-cfclient"
+	"testing"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
 
-var _ = Describe("Cloud Foundry library code", func() {
-	Context("With a Fake Cloud Foundry", func() {
-		var (
-			cf     CloudFoundry
-			client *cfclient.Client
-		)
-
-		BeforeEach(func() {
-			cf = NewFakeCloudFoundry()
-
-			var err error
-			client, err = cfclient.NewClient(&cfclient.Config{
-				ApiAddress: cf.Server.URL,
-				Token:      "foobar",
-			})
-
-			Expect(err).To(BeNil())
-		})
-
-		AfterEach(func() {
-			cf.Teardown()
-		})
-
-		It("should not have any apps", func() {
-			apps, err := client.ListApps()
-			Expect(err).To(BeNil())
-
-			Expect(apps).To(BeEquivalentTo([]cfclient.App{}))
-		})
-	})
-})
+func TestConfig(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "watchful internal config suite")
+}
